requirements.txt
ollama
pydub
scenedetect
ffmpeg-python
opencv-python
edge-tts
Pillow
python-dotenv
requests